Cadastro de Pessoa

+++ src/main/java/com/roderly/microbiomelabufu/endereco/dto/request/EnderecoRequest.java
@@ -7,7 +7,7 @@ public record EnderecoRequest(
         String complemento,
         String bairro,
         String cidade,
-        Integer id_uf,
+        Long id_uf,
         String cep,
         String referencia
 ){}
