+++ src/main/java/com/roderly/microbiomelabufu/pessoa/model/Pessoa.java
@@ -19,13 +19,13 @@ public class Pessoa {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id_pessoa;
 
-    private Integer id_tipo_pessoa;
+    private Long id_tipo_pessoa;
 
     private String nome;
 
     private String sobrenome;
 
-    private Integer id_sexo;
+    private Long id_sexo;
 
     private LocalDate data_nascimento;
 
@@ -33,18 +33,17 @@ public class Pessoa {
 
     private String email;
 
-    private Integer id_estado_civil;
+    private Long id_estado_civil;
 
-    private Integer id_profissao;
+    private Long id_profissao;
 
     private String cpf;
 
-    private Integer id_etnia;
+    private Long id_etnia;
 
-    private Integer id_escolaridade;
+    private Long id_escolaridade;
 
-    private Integer id_religiao;
-
+    private Long id_religiao;
 
     @OneToMany(mappedBy = "pessoa", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Endereco> endereco = new ArrayList<>();
